dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_Range-detach.html

<!DOCTYPE html>
<title>Range.detach</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(function() {
  var r = document.createRange();
  r.detach()
  assert_throws("INVALID_STATE_ERR", function() { r.detach() })
})
</script>
